+++ server.cpp
@@ -51,7 +51,7 @@ int main(int argc, char *argv[])
 	//setup random number generator for the log relation key (a random number that related logs can use)
 	random_device rd;
 	mt19937 mt(rd());
-	uniform_int_distribution<unsigned long> dist (0, 9223372036854775807);
+	uniform_int_distribution<unsigned long> dist (0, (unsigned long)9223372036854775807);
 	unsigned long initkey = dist(mt);
 
 	//you MUST establish the postgres utilities instance variable here or get a segmentation inside on c->prepare
